backend/pyproject.toml
[project]
name = "diffcast-backend"
version = "0.1.0"
description = "Diffcast - GitHub diff to video pipeline"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "firebase-admin>=6.0.0",
    "python-dateutil>=2.8.0",
    "google-genai>=1.0.0",
    "browser-use[video]==0.11.9",
    "playwright>=1.58.0",
    "langchain-google-genai>=4.2.0",
    "google-cloud-texttospeech>=2.16.0",
    "Pillow>=10.0.0",
]
